refactor: move all models to new `models` module
